add styles

+++ src/components/Contact/Contact.jsx
@@ -7,6 +7,7 @@ import s from './Contact.module.css';
 import { useDispatch } from 'react-redux';
 import { onEdit } from '../../redux/editSlice';
 import { deleteContactsThunk } from '../../redux/contacts/operations';
+import Swal from 'sweetalert2';
 
 const Contact = ({ id, name, number }) => {
   const dispatch = useDispatch();
@@ -37,7 +38,24 @@ const Contact = ({ id, name, number }) => {
         className={clsx(s.signUp)}
         type="button"
         onClick={() => {
-          dispatch(deleteContactsThunk(id));
+          Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+          }).then(result => {
+            if (result.isConfirmed) {
+              dispatch(deleteContactsThunk(id));
+              Swal.fire({
+                title: 'Deleted!',
+                text: 'Your file has been deleted.',
+                icon: 'success',
+              });
+            }
+          });
         }}
       >
         Delete
